[FIX] hr_attendance: display of attendance button in systray
In the enterprise client, in mobile, the buttons in the systray are 'display:
inline-block'. This rule was overriden in the case of the attendance button
by a 'this.$el.show()', which changed the display to 'block'. This rev. uses
'do_show' and 'do_hide' functions of widgets instead, as they don't change the
display.