[FIX] account: allow forcing company when fetching fpos
The code to fetch the fiscal position of a partner uses ACLs and
ir_rules to select the correct ones, but when calling the function
as admin, multi-company rules flight out the window.

By allowing to force a company in the context, we can get the fpos
as the admin from the pov of any company.