[FIX] account: partner_id field in account.move
From 94ed4ce8c8354b432d9d5c90c2f4631347fb4398

The original issue is the following:
- Go to the journal item creation
- Choose a partner
- Add a line => the partner on the journal item is removed

This is because `partner_id` on the account move is a related to
`partner_id` on the account move lines. This is likely to cause errors
during the manual encoding of a journal item, since the relation will
link the partner of the account move to the partner of the first account
move line only.

opw:672354