[FIX] account: create_analytic_lines
When deleting the analytic account of an account move line(aml),
the analytic lines linked to this aml must be unlinked even if
the aml is not linked to another account analytic.

Used case:

Once a analytic account is removed from a journal line, the analytic line
created for this journal line has to be removed from this analytic account.
Inspired from 8.0

opw:673803