[IMP] account: short code always required
The short code is always required for the following reasons:
- The auto-generated code is useless for search purpose because user
  cannot choose it (BNK or CSH is maybe not meaningful for him)
- If the system cannot generate an unused journal code, the user is
  stuck since the field is read-only in the interface.

However, the code can still be auto-generated, although it is required
in the interface. This is to avoid breaking modules that would expect
this feature. At the same time, the logic has been improved to reduce
the number of search requests.

opw-674485