[ADD] base: view inheritance with wrapping support
When using the inheritance spec `position="replace"`,
the framework will now support a text node exactly
equal to `$0` and will replace it with the node
that was matched by the inheritance locator.

Given an original view equal to this:
```
   <div class="foo">
     <p class="bar">the paragraph</p>
   </div>
```

And an inherited view with the following inheritance spec:
```
   <p class="bar" position="replace">
      <div class="panel">
        <div class="panel-header">The Title</div>
          <div class="panel-body">
            $0
          </div>
      </div>
   </p>
```

The finalt result after inheritance would become:
```
   <div class="foo">
      <div class="panel">
        <div class="panel-header">The Title</div>
        <div class="panel-body">
          <p class="bar">the paragraph</p>
        </div>
      </div>
   </div>
```

Also updated doc and tests accordingly.