[IMP] mail: mention: don't delay local mention lookups
A 200ms delay is set on lookups for mention suggestions to prevent triggering
RPCs when the user is still typing. However, there shouldn't be a delay when
the suggestions are looked up in the prefetched suggestions.

Also fixes a bug that might occur when the user starts typing a mention which
triggers an RPC, but erase it before the RPC returns.