Revert "[FIX] res.currency: allow duplication"
This reverts commit 1d1df8013b2184c16089793ee3905eb27c10407e.

Did not work due to the size=3 on name field that strips the code to the first
three letters only (removing the " (copy)" part).
Copying a currency has a few business cases as the rates are not copied.
As can not increase the size of a field in stable, remove the method that had no
effect.

Fixes #11036