[FIX] web,base: m2o empty name results
M2O use the autocomplete lib to display its content. The problem is
that if a result is empty (or only composed of whitespaces), a
divider is displayed in the dropdown.

Commit d627068c0f0249f8bd6608ff70ef702a60540c16 was fixing this by only displaying non-empty
name results... but it did not work as many models implement the
name_search method the wrong way.

The new solution is to add a default display for record values which
are empty strings, at the right places in the webclient components.