[IMP] website_payment: when invoicing always ignore invoice policy
Sale orders created through the website that have to be automatically
invoiced should ignore the delivery invoicing policy.

Otherwise you can end up with either an invoice not being generated at
all. This means the user has to manually generate and add the payment to
it. Even worse is if only some purchased products have invoice policy
delivery. In this case a partial invoice would be generated to which the
payment would be added, leading to a mismatch between invoice total and
payment amount.