[FIX] web: data_manager: load_action: return deep copy of action
The action object, and in particular its context, is modified by the Action
Manager and by some views (e.g. sales_team_dashboard Kanban). So, as the action
is stored in cache, the same object was returned each time the action was
requested, and previous modifications of its context still applied.

This produced a bug with the sales_team_dashboard when following those steps:
  - click on the link to today's activities, it opens an action with
    'search_default_today' in the context,
  - come back to the dashboard
  - click on the link to this week's activities, it opens an action with
    'search_default_this_week' in the context, but the formerly added key
    'search_default_today' is still there as well, so the domain is wrong.