[IMP] sales_team: don't auto install sales_cteam
Module sales_team was auto installed if base and mail was installed, but
this isn't wanted as it is already a depends of crm and sale modules
which should be anyway installed for sales team to have a meaning.

This could lead to a glitch with the sale apps displayed whilst it
should not be.

opw-679223